requirements.txt
google-generativeai
spacy # Name entity recognition
streamlit
nltk
wordcloud
python-dotenv
pymupdf
beautifulsoup4
requests


-e .







